+++ src/_middleware/lifecycleHandlers.ts
@@ -15,6 +15,6 @@ export const checkMaintenanceMode = ({set, store}:any) => {
   if(store.maintenanceMode){
       console.log("Maintenance Mode: ON");
       set.status = HttpStatusEnum.HTTP_423_LOCKED;
-      return { message: 'Server currently undergoing maintenance' };
+      return { message: 'System currently undergoing maintenance', note: 'Admin has locked this application' };
   }
 }